% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{addHalfInterval}
\alias{addHalfInterval}
\title{Calculate mid-classes of an age interval}
\usage{
addHalfInterval(x)
}
\arguments{
\item{x}{Vector with beginning of age classes.}
}
\description{
Calculate mid-classes of an age interval. \cr\cr
\emph{\bold{Internal function}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
