% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{age2dx}
\alias{age2dx}
\title{Constructing \code{dx} from individual ages/times at deaths}
\usage{
age2dx(times, x)
}
\arguments{
\item{times}{vector with age/times at deaths.}

\item{x}{Beginning of the Age/time class. The resulting vector of \code{dx} will have the same length.}
}
\description{
The function converts ages/times at death into \code{dx} according.
Notice that age at deaths are assumed to be exactly observed.\cr\cr
\emph{\bold{Internal function}}
}
\examples{
\dontrun{
# Data:
x=c(0,0.5,2,5,10,13,15)
dx=c(1,2,6,15,22,6,1)

# Should give the same vector as in dx.
age2dx(dx2age(dx,x),x)

}
}
\seealso{
\code{\link{lx2dx}}, \code{\link{dx2lx}}, and \code{\link{dx2age}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
