% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{all.elements.equal}
\alias{all.elements.equal}
\title{Checking if all elements in the vector are equal}
\usage{
\method{all}{elements.equal}(x)
}
\arguments{
\item{x}{A vector to be analyzed.}
}
\description{
Checking if all elements in the vector are equal. .\cr\cr
\emph{\bold{Internal function}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
