% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{Bootstrap_dx}
\alias{Bootstrap_dx}
\title{Fast Bootstrap method performed on \code{dx}}
\usage{
Bootstrap_dx(dx, x, pash.parent, N = 1000, pace.type = "all",
  shape.type = "all", q = 0.5, harmonized = TRUE)
}
\arguments{
\item{dx}{A vector with integer counts (deaths).}

\item{x}{Beginning of the Age/time class.}

\item{pash.parent}{A parent pash object.}

\item{N}{Number of bootstrap replicates.}

\item{pace.type}{Which pace measure should be returned (default "all")?
Use "none" if you don't wat to return any pace measure. See \code{\link{GetPace}} for details.}

\item{shape.type}{Which shape measure should be returned (default "all")?
Use "none" if you don't wat to return any shape measure. See \code{\link{GetShape}} for details.}

\item{q}{GetPace parameter. Quantile specification for age where q percent of the life-table population is still alive (defaults to median).
See \code{\link{GetPace}} for details.}

\item{harmonized}{GetShape parameter. Should the harmonized version of the shape measures be returned (default \code{TRUE})?
See \code{\link{GetShape}} for details.}
}
\description{
Fast Bootstrap method performed on \code{dx}.\cr\cr
\emph{\bold{Internal function}}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
