% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{confint.pash}
\alias{confint.pash}
\title{Claculates condfidence intervals for pash measures}
\usage{
\method{confint}{pash}(object, population.size, N = 1000, bs.er = 0.25,
  jk.er = 0.25, pace.type = "all", shape.type = "all", q = 0.5,
  harmonized = TRUE, trace = TRUE, alpha = 0.05)
}
\arguments{
\item{object}{\code{Pash} object.}

\item{population.size}{The size of population for the constructed pash object. If not given then it will be read from the \code{object}.}

\item{N}{Number of bootstrap replicates.}

\item{pace.type}{Which pace measure should be returned (default "all")?
Use "none" if you don't wat to return any pace measure. See \code{\link{GetPace}} for details.}

\item{shape.type}{Which shape measure should be returned (default "all")?
Use "none" if you don't wat to return any shape measure. See \code{\link{GetShape}} for details.}

\item{q}{GetPace parameter. Quantile specification for age where q percent of the life-table population is still alive (defaults to median).
See \code{\link{GetPace}} for details.}

\item{harmonized}{GetShape parameter. Should the harmonized version of the shape measures be returned (default \code{TRUE})?
See \code{\link{GetShape}} for details.}
}
\description{
Claculates condfidence intervals around pash measures for a given \code{pash} object. \cr\cr Generic function.
}
\examples{
\dontrun{

obj <- Inputlx(x = australia_10y$x, lx = australia_10y$lx,nax = australia_10y$nax, nx = australia_10y$nx, last_open = TRUE)

z <- confint(object = obj, population.size = 1000, trace = TRUE)

print(z, CI.type = 'BCa')
print(z, CI.type = 'Percentile')

}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall.
}
