% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{dx2age}
\alias{dx2age}
\title{Constructing vector of age at deaths assuming that each death occurs in the middle of an age interval}
\usage{
dx2age(dx, x)
}
\arguments{
\item{dx}{Death counts vector.}

\item{x}{Beginning of the Age/time class. A vector of the same length as \code{ndx}.}
}
\description{
The function converts counts in vector dx into into individual age/time at deaths.
Notice that dx does not contain information about censoring so resulting vector has only exactly observed events.\cr\cr
\emph{\bold{Internal function}}
}
\examples{
\dontrun{
# Data:
x=c(0,0.5,2,5,10,13,15)
dx=c(1,2,6,15,22,6,1)

# Should give the same vector as in dx.
age2dx(dx2age(dx,x),x)

}
}
\seealso{
\code{\link{lx2dx}}, \code{\link{dx2lx}}, and \code{\link{age2dx}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
