% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{dx2lx}
\alias{dx2lx}
\title{Converting dx to lx assuming that all events are exactly observed (no censoring)}
\usage{
dx2lx(dx)
}
\arguments{
\item{dx}{Vector with death counts}
}
\value{
A vector with population size at the beginning of interval.
}
\description{
Converting dx to lx assuming that all events are exactly observed (no censoring). \cr\cr
\emph{\bold{Internal function}}
}
\details{
The presence or absence of open interval does not matter here.
}
\seealso{
\code{\link{lx2dx}}, \code{\link{dx2age}}, and \code{\link{age2dx}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
