% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{getInterval}
\alias{getInterval}
\title{Calculate midd-classes of an age interval}
\usage{
getInterval(x)
}
\arguments{
\item{x}{Vector with begining of age classes.}
}
\description{
Calculate midd-classes of an age interval. \cr\cr
\emph{\bold{Internal function}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
