% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{getpash}
\alias{getpash}
\title{Calculating pace and shape measures as one vector for bootstrap computations}
\usage{
getpash(dx, x, pash.parent, pace.type = "all", shape.type = "all",
  q = 0.5, harmonized = TRUE)
}
\arguments{
\item{dx}{A vector with integer counts (deaths).}

\item{x}{Beginning of the Age/time classes. Vector with the same length as \code{dx}}

\item{pash.parent}{A parent pash object.}

\item{pace.type}{Which pace measure should be returned (default "all")?
Use "none" if you don't want to return any pace measure. See \code{\link{GetPace}} for details.}

\item{shape.type}{Which shape measure should be returned (default "all")?
Use "none" if you don't want to return any shape measure. See \code{\link{GetShape}} for details.}

\item{q}{GetPace parameter. Quantile specification for age where q percent of the life-table population is still alive (defaults to median).
See \code{\link{GetPace}} for details.}

\item{harmonized}{GetShape parameter. Should the harmonized version of the shape measures be returned (default \code{TRUE})?
See \code{\link{GetShape}} for details.}
}
\description{
Calculating pace and shape measures as one vector for bootstrap computations. \cr\cr
\emph{\bold{Internal function}}
}
\keyword{internal}
