% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{JackKnife_dx}
\alias{JackKnife_dx}
\title{Fast JackKnife method performed on \code{dx}}
\usage{
JackKnife_dx(dx, x, ...)
}
\arguments{
\item{dx}{A vector with integer counts (deaths).}

\item{x}{Beginning of the Age/time classes. Vector with the same length as \code{dx}.}

\item{...}{Other parameters passed to \code{\link{getpash}} function.}
}
\description{
Fast JackKnife method performed on \code{dx}.\cr\cr
\emph{\bold{Internal function}}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
