% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{JackKnifeAcc}
\alias{JackKnifeAcc}
\title{Calculation of JeckKnife acceleration parameter for BCA method}
\usage{
JackKnifeAcc(JackKnifeMat, dx)
}
\arguments{
\item{JackKnifeMat}{JackKnife matrix with one row per each considered pash measure. See \code{\link{JackKnife_dx}}.}

\item{dx}{A vector with integer counts (deaths).}
}
\description{
Calculation of JeckKnife acceleration parameter for BCA method. \cr\cr
\emph{\bold{Internal function}}
}
\examples{
\dontrun{

#Get some data
population.size <- 10000
obj <- Inputlx(x = australia_10y$x, lx = australia_10y$lx,nax = australia_10y$nax, nx = australia_10y$nx, last_open = TRUE)
dx <- lx2dx(round(object$lt$lx*population.size))
x <- obj$lt$x

#Acceleration parameters calculated by slow method:
times <- dx2age(dx,x)
gp <- function(times, x, pash.parent) getpash(dx = age2dx(times = times, x=x), x=x, pash.parent = pash.parent)
J1 <- as.matrix(sapply(seq_along(times), function(k) gp(times = times[-k], x = x, pash.parent = obj)))
a.slow <- JackKnifeAcc(J1)

#Acceleration parameters calculated by fast method:
J2 <- JackKnife_dx(dx = org.dx, x = org.x, pash.parent = obj)
a.fast <- JackKnifeAcc(J2,dx)

#Compare results
cbind(a.slow = a.slow, a.fast = a.fast, diff = round(a.slow - a.fast, 12))

}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall. Page 186.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
