% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{lx2dx}
\alias{lx2dx}
\title{Converting lx to dx assuming that all events are exactly observed (no censoring)}
\usage{
lx2dx(lx)
}
\arguments{
\item{lx}{population size at the beginning of the age interval.}
}
\value{
A vector with death counts.
}
\description{
Converting lx to dx assuming that all events are exactly observed (no censoring). \cr\cr
\emph{\bold{Internal function}}
}
\seealso{
\code{\link{dx2lx}}, \code{\link{dx2age}}, and \code{\link{age2dx}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
