% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{PER.CI}
\alias{PER.CI}
\title{Calculation of the Percentile type of confidence intervals}
\usage{
PER.CI(OrgEst, BootEst, alpha = 0.05)
}
\description{
Calculation of the Percentile type of confidence intervals. \cr\cr
\emph{\bold{Internal function}}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
