% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{BCA.CI}
\alias{BCA.CI}
\title{Calculation of the BCA type of confidence intervals}
\usage{
BCA.CI(OrgEst, BootEst, JackKnifeMat, Orgdx, alpha = 0.05)
}
\arguments{
\item{OrgEst}{Output of \code{\link{getpash}} for original dataset.}

\item{BootEst}{Output of \code{\link{getpash}} for all bootstrap replicates.}

\item{JackKnifeMat}{Output of \code{\link{getpash}} for JackKnife samples. See \code{\link{JackKnife_dx}}.}

\item{Orgdx}{Vector with counts (deaths) for original dataset.}

\item{alpha}{Significance level.}
}
\description{
Calculation of the BCA type of confidence intervals. \cr\cr
\emph{\bold{Internal function}}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
