% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{confint.pash}
\alias{confint.pash}
\title{Calculating confidence intervals for pash measures}
\usage{
\method{confint}{pash}(object, population.size, N = 1000, pace.type = "all",
  shape.type = "all", q = 0.5, harmonized = TRUE, trace = TRUE,
  alpha = 0.05, bs.er = 0.25, jk.er = 0.1)
}
\arguments{
\item{object}{\code{Pash} object.}

\item{population.size}{The size of population for the constructed \code{pash} object. If not given then it will be read from the \code{object}.}

\item{N}{Number of bootstrap replicates.}

\item{pace.type}{Which pace measure should be returned (default "all")?
Use "none" if you don't wat to return any pace measure. See \code{\link{GetPace}} for details.}

\item{shape.type}{Which shape measure should be returned (default "all")?
Use "none" if you don't wat to return any shape measure. See \code{\link{GetShape}} for details.}

\item{q}{GetPace parameter. Quantile specification for age where q percent of the life-table population is still alive (defaults to median).
See \code{\link{GetPace}} for details.}

\item{harmonized}{GetShape parameter. Should the harmonized version of the shape measures be returned (default \code{TRUE})?
See \code{\link{GetShape}} for details.}

\item{trace}{Logical indicating if to show summary of performed bootstrap.}

\item{alpha}{Significance level.}

\item{bs.er}{Maximal acceptable fraction of Bootstrap errors 
(a measure does not exists for a certain bootstrapped data).}

\item{js.er}{Maximal acceptable fraction of JackKnife errors 
(a measure does not exists for a certain JackKnife cases).}
}
\description{
Calculating confidence intervals around pash measures for a given \code{pash} object. \cr\cr Generic function.
}
\examples{
\dontrun{

#Get a data
obj <- Inputlx(x = australia_10y$x, lx = australia_10y$lx,nax = australia_10y$nax, nx = australia_10y$nx, last_open = TRUE)

ci1 <- confint(object = obj, population.size = 1000, trace = TRUE)
print(ci1, CI.type = 'BCa')
print(ci1, CI.type = 'Percentile')

#Small population size
obj <- Inputlx(x = australia_10y$x, lx = australia_10y$lx,nax = australia_10y$nax, last_open = FALSE) 
ci2 <- confint(object = obj, population.size = 300, trace = TRUE)
ci2

#Very small population size
# There is a mistake in pash package. The linear extrapolation doesn't work for some sets with open last intervals.
obj <- Inputlx(x = australia_10y$x, lx = australia_10y$lx,nax = australia_10y$nax, nx = australia_10y$nx, last_open = TRUE)
ci3 <- confint(object = obj, population.size = 10, shape.type = 'none', trace = TRUE)

}
}
\references{
Efron, B., & Tibshirani, R. J. (1993). An introduction to the bootstrap. New York: Chapman & Hall.
}
