% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpash.r
\name{print.bootpash}
\alias{print.bootpash}
\title{Printing the confidence intervals for pash object}
\usage{
\method{print}{bootpash}(object, CI.type = c("BCa", "Percentile"),
  digits = 4)
}
\arguments{
\item{CI.type}{The type of printed confidence intervals. One from "BCa" or "Percentile".}

\item{oject}{An \code{bootpash} object with fitted confidence interval for \code{pash} object.}
}
\description{
Generic function to print the confidence intervals for pash object.
}
